requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
firebase-admin==6.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
django-stubs==4.2.7
django-cors-headers==4.9.0